[FIX] mail: fix activity state is shared accross tabs
Since [1], the activity state, which is supposed to be shared
accross tab through a broadcast channel, isn't anymore.

This PR fixes the responsible typo.

[1]: #161286

X-original-commit: b81ec33f22ee0361fffa30560bab4ed783d8fda4

+++ addons/mail/static/src/core/web/activity_model.js
@@ -201,7 +201,7 @@ export class Activity extends Record {
             [[this.id]],
             { feedback: this.feedback }
         );
-        this.activityBroadcastChannel?.postMessage({
+        this.store.activityBroadcastChannel?.postMessage({
             type: "RELOAD_CHATTER",
             payload: { id: this.res_id, model: this.res_model },
         });
@@ -209,9 +209,12 @@ export class Activity extends Record {
     }
 
     remove({ broadcast = true } = {}) {
+        if (!this.exists()) {
+            return;
+        }
         this.delete();
         if (broadcast) {
-            this.activityBroadcastChannel?.postMessage({
+            this.store.activityBroadcastChannel?.postMessage({
                 type: "DELETE",
                 payload: { id: this.id },
             });
